[UPDATE]: run-tsk was updated

+++ src/run-tsk/src/install.sh
@@ -7,11 +7,19 @@ fi
 # Print the current directory
 echo "Working on directory: $PWD"
 
+# Check if pnpm or npm is installed in the system
+PNPM_VERSION=$(pnpm --version 2>/dev/null)
+NPM_VERSION=$(npm --version 2>/dev/null)
+if [ -z "$PNPM_VERSION" ] && [ -z "$NPM_VERSION" ]; then
+  echo "Error: pnpm or npm is required to install the dependencies"
+  exit 1
+fi
+
 # Clone the repository from https://github.com/harvic3/nodetskeleton.git
 echo "Cloning the nodetskeleton from origin"
 git clone https://github.com/harvic3/nodetskeleton.git
 
-# rename the current directory name nodetskeleton to the project-name provided
+# Rename the current directory name nodetskeleton to the project-name provided
 echo "Preparing your project $1"
 mv nodetskeleton $1
 rm -rf nodetskeleton
@@ -19,8 +27,13 @@ rm -rf nodetskeleton
 # Install dependencies
 echo "Installing dependencies"
 cd $1
-pnpm --version || npm install -g pnpm
-pnpm install
+if [ ! -z "$PNPM_VERSION" ]; then
+  echo "Using pnpm $PNPM_VERSION"
+  pnpm install
+else
+  echo "Using npm $NPM_VERSION"
+  npm install
+fi
 
 # Create resources
 echo "NODE_ENV=development
@@ -32,8 +45,10 @@ ORIGINS=http://localhost:3003
 ENCRYPTION_KEY=JUS9192ZliRlDBWm0BmmJoZO1PbNkZt3kiXNlaGLkIT49uEdgGe79TPCbr0D
 ENCRYPTION_ITERATIONS=4e4
 ENCRYPTION_KEY_SIZE=128
-JWT_SECRET_KEY=2NtC29d33z1AF1HdPSpn" > .env
+JWT_SECRET_KEY=2NtC29d33z1AF1HdPSpn
+JWT_EXPIRE_IN_SECONDS=3600" > .env
 
+# Finalizing the installation
 echo "Your project $1 is ready"
 echo "Now go to the project directory typing 'cd $1'"
 echo " And type 'npm run dev' to start the server"
